% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisPlot.R
\name{box_plot}
\alias{box_plot}
\title{Box Plot}
\usage{
box_plot(
  data = NULL,
  x,
  y,
  xlab = "x-axis",
  ylab = "y-axis",
  title = "Box Plot"
)
}
\arguments{
\item{data}{A data.frame or any other object provided will be used to replace
the default plot data. All objects will be transformed into a fortified
data.frame. You can refer to the fortify() function to understand which
variables will be created. Alternatively you can use a custom function that
takes the plot data as a single argument. The function must return a
data.frame, which will be utilized as the layer data.}

\item{x}{The "x" parameter specifies the variable or the data that will be
used to create the box plot. It represents the data that will be grouped or
split into different categories to visualize their distribution and
variability.}

\item{y}{The "y" parameter specifies the numerical variable that will be
represented on the y-axis of the plot. It is the variable for which the
distribution and variability will be visualized using the box plot.}

\item{xlab}{The "xlab" parameter allows you to specify the label or title for
the x-axis in your plot. The x-axis label is essential for effectively
communicating the meaning of the data represented on the horizontal axis.}

\item{ylab}{The "ylab" parameter allows you to specify the label or title for
the y-axis in your plot. The y-axis label is crucial for effectively
communicating the meaning of the data represented on the vertical axis.}

\item{title}{The "title" parameter is used to specify the title of the plot.
The title provides a clear and concise description of the plot's purpose or
the main idea it conveys. It is typically placed above the plot to provide
context and help users understand the data being visualized.}

\item{fill}{The "fill" parameter is used to specify the variable or data that
will be used to fill the boxes. It allows you to color the boxes based on a
categorical variable, providing visual distinction between different groups
or categories of data.}
}
\description{
Creates a box plot using ggplot2.
Examine the distribution of data and identify potential outliers and
variability among different categories or groups. The boxplot is particularly
valuable for comparing data distributions side by side.
}
\examples{
box_plot(data = information_full, x = alter, y = interesse,
            xlab = "age", ylab = "interest",
            title = "correlation of age and interests"),
            geom_boxplot(fill = "lightblue"))
}
