% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisPlot.R
\docType{data}
\name{colours}
\alias{colours}
\title{Colour}
\format{
An object of class \code{character} of length 11.
}
\usage{
colours
}
\arguments{
\item{colour}{The colour for the points.}
}
\description{
Define farbcodes, so it will be easier to choose colours.
}
\keyword{datasets}
