% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisPlot.R
\name{jitter_plot}
\alias{jitter_plot}
\title{Jitter plot}
\usage{
jitter_plot(
  data = NULL,
  x,
  y,
  xlab = "x-axis",
  ylab = "y-axis",
  colour = NULL,
  title = "Jitter Plot"
)
}
\arguments{
\item{data}{A data.frame or any other object provided will be used to replace
the default plot data. All objects will be transformed into a fortified
data.frame. You can refer to the fortify() function to understand which
variables will be created. Alternatively you can use a custom function that
takes the plot data as a single argument. The function must return a
data.frame, which will be utilized as the layer data.}

\item{x}{The "x" parameter would define the variable that will be plotted on
the horizontal axis.}

\item{y}{The "y" parameter would define the variable that will be plotted on
the vertical axis.}

\item{xlab}{The "xlab" parameter allows you to specify the label or title for
the x-axis in your plot. The x-axis label is essential for effectively
communicating the meaning of the data represented on the horizontal axis.}

\item{ylab}{The label for the y-axis.}

\item{colour}{The colour for the jitters.}

\item{title}{The title of the plot.}
}
\description{
Creates a jitter plot using ggplot2.
Uncover the distribution of data points along a single axis, revealing the
density and spread of values, particularly useful for smaller datasets.
}
\examples{

jitter_plot(data = christmas_movies_clean, x = release_year, y = stars,
            xlab = "release year", ylab = "number of evaluations per year",
            colour = "lightgreen",
            title = "number of evaluations per year through release years")
}
