% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisPlot.R
\name{jitter_plot_shape}
\alias{jitter_plot_shape}
\title{Jitter plot with different shapes}
\usage{
jitter_plot_shape(
  data = NULL,
  x,
  y,
  xlab = "x-axis",
  ylab = "y-axis",
  shape = NULL,
  colour = NULL,
  title = "Jitter plot with point shapes"
)
}
\arguments{
\item{data}{A data.frame or any other object provided will be used to replace
the default plot data. All objects will be transformed into a fortified
data.frame. You can refer to the fortify() function to understand which
variables will be created. Alternatively you can use a custom function that
takes the plot data as a single argument. The function must return a
data.frame, which will be utilized as the layer data.}

\item{x}{The variable to be shown on the x-axis.}

\item{y}{The variable to be shown on the y-axis.}

\item{xlab}{The label for the x-axis.}

\item{ylab}{The label for the y-axis.}

\item{shape}{The shape for the points.}

\item{colour}{The colour for the points.}

\item{title}{The title of the plot.}

\item{size}{The size of the points.}
}
\description{
Creates a jitter plot with different shapes using ggplot2.
}
\examples{
jitter_plot_shape(data = oktoberfest_daten, x = bier_preis, y = bier_konsum,
            xlab = "beer price", ylab = "beer consumption",
            shape = "circle", colour = "orange",
            title = "Relationship between beer consumption and beer price",
            subtitle = "depending on the number of visitors per day")
}
