% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisPlot.R
\name{jitterplot_shape}
\alias{jitterplot_shape}
\title{Jitter Plot with point shapes (ggplot2)}
\usage{
jitterplot_shape(
  data,
  x,
  y,
  xlab = "x-axis",
  ylab = "y-axis",
  shape = "circle",
  colour = "orange",
  title = "Jitter Plot with point shapes"
)
}
\arguments{
\item{data}{The data frame who contains the data.}

\item{x}{The variable to be seen on the x-axis.}

\item{y}{The variable to be seen on the y-axis.}

\item{xlab}{The label for the x-axis.}

\item{ylab}{The label for the y-axis.}

\item{shape}{The shape for the points.}

\item{colour}{The colour for the points.}

\item{title}{The title of the plot.}

\item{size}{The size of the points.}
}
\description{
Jitter Plot with point shapes (ggplot2)
}
