% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisPlot.R
\name{line_chart}
\alias{line_chart}
\title{Line chart}
\usage{
line_chart(
  data = NULL,
  x,
  y,
  xlab = "x-axis",
  ylab = "y-axis",
  colour = "blue",
  title = "Line chart"
)
}
\arguments{
\item{data}{A data.frame or any other object provided will be used to replace
the default plot data. All objects will be transformed into a fortified
data.frame. You can refer to the fortify() function to understand which
variables will be created. Alternatively you can use a custom function that
takes the plot data as a single argument. The function must return a
data.frame, which will be utilized as the layer data.}

\item{x}{The "x" parameter would typically represent time or another continuous
variable that is used to establish the order of the data points.}

\item{y}{The variable to be shown on the y-axis.}

\item{xlab}{The "xlab" parameter allows you to specify the label or title for
the x-axis in your plot. The x-axis label is essential for effectively
communicating the meaning of the data represented on the horizontal axis.}

\item{ylab}{The label for the y-axis.}

\item{colour}{The colour for the line.}

\item{title}{The title of the plot.}
}
\description{
Creates a line chart using ggplot2.
Visualise trends and changes in data over continuous intervals, perfect for
time-series data or any dataset with a natural ordering.
}
\examples{
line_chart(data = oktoberfest_daten, x = jahr, y = bier_konsum,
            xlab = "year", ylab = "beer consumption (in hl)",
            colour = "blue",
            title = "beer consumption during the Oktoberfest")
}
