% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisPlot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Simplifize the visualisation of plots. In the past two semester we did have to
plot many plots and we thought that it would be easier, if users would know
exactly how they have to plot something. The package visplots is there to
minimalize the work to write a good plot.}
\usage{
scatter_plot(
  data,
  x,
  y,
  xlab = "x-axis",
  ylab = "y-axis",
  colour = "purple",
  size = 5,
  title = "Scatter Plot"
)
}
\arguments{
\item{data}{The data frame who contains the data.}

\item{x}{The variable to be shown on the x-axis.}

\item{y}{The variable to be shown on the y-axis.}

\item{xlab}{The label for the x-axis.}

\item{ylab}{The label for the y-axis.}

\item{colour}{The colour for the points.}

\item{size}{The size of the points.}

\item{title}{The title of the plot.}
}
\description{
Creates a scatter plot using ggplot2.
Discover relationships and patterns between two numerical variables with a
scatter plot, ideal for understanding correlations and identifying outliers.
}
\examples{
scatter_plot(data = oktoberfest_daten, x = jahr, y = besucher_gesamt,
            xlab = "year", ylab = "number of visitors (in millions)",
            colour = "purple", size = 3,
            title = "number of vistors of the Oktoberfest per year")
}
