% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisPlot.R
\name{scatterplot}
\alias{scatterplot}
\title{Scatterplot (ggplot2)}
\usage{
scatterplot(
  data,
  x,
  y,
  xlab = "x-axis",
  ylab = "y-axis",
  colour = "purple",
  size = 5,
  title = "Scatterplot"
)
}
\arguments{
\item{data}{The data frame who contains the data.}

\item{x}{The variable to be seen on the x-axis.}

\item{y}{The variable to be seen on the y-axis.}

\item{xlab}{The label for the x-axis.}

\item{ylab}{The label for the y-axis.}

\item{colour}{The colour for the points.}

\item{size}{The size of the points.}

\item{title}{The title of the plot.}
}
\description{
Scatterplot (ggplot2)
}
