% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisPlot.R
\name{jitter_plot}
\alias{jitter_plot}
\title{Jitter Plot (ggplot2)}
\usage{
jitter_plot(
  data,
  x,
  y,
  xlab = "x-axis",
  ylab = "y-axis",
  colour = "lightgreen",
  title = "Jitter Plot"
)
}
\arguments{
\item{data}{The data frame who contains the data.}

\item{x}{The variable to be shown on the x-axis.}

\item{y}{The variable to be shown on the y-axis.}

\item{xlab}{The label for the x-axis.}

\item{ylab}{The label for the y-axis.}

\item{colour}{The colour for the jitters.}

\item{title}{The title of the plot.}
}
\description{
Jitter Plot (ggplot2)
}
