% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisPlot.R
\name{jitter_plot_shape}
\alias{jitter_plot_shape}
\title{Creates a jitter plot with different shapes using ggplot2.}
\usage{
jitter_plot_shape(
  data,
  x,
  y,
  xlab = "x-axis",
  ylab = "y-axis",
  shape = "circle",
  colour = "orange",
  title = "Jitter plot with point shapes"
)
}
\arguments{
\item{data}{The data frame who contains the data.}

\item{x}{The variable to be shown on the x-axis.}

\item{y}{The variable to be shown on the y-axis.}

\item{xlab}{The label for the x-axis.}

\item{ylab}{The label for the y-axis.}

\item{shape}{The shape for the points.}

\item{colour}{The colour for the points.}

\item{title}{The title of the plot.}

\item{size}{The size of the points.}
}
\description{
Creates a jitter plot with different shapes using ggplot2.
}
\examples{
jitter_plot_shape(data = oktoberfest_daten, x = bier_preis, y = bier_konsum,
            xlab = "price for beer", ylab = "beer consumption", shape = "circle",
            colour = "orange",
            title = "Relationship between beer consumption and beer price",
            subtitle = "depending on the number of visitors per day")
}
