% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisPlot.R
\name{jitter_plot_shape}
\alias{jitter_plot_shape}
\title{Jitter Plot with different shapes}
\usage{
jitter_plot_shape(
  data = NULL,
  x,
  y,
  xlab = "x-axis",
  ylab = "y-axis",
  shape = NULL,
  colour = NULL,
  title = "Jitter Plot with different shapes"
)
}
\arguments{
\item{data}{A data.frame or any other object provided will be used to replace
the default plot data. Alternatively you can use a custom function that
takes the plot data as a single argument. The function must return a
data.frame, which will be utilised as the layer data.}

\item{x}{The "x" parameter would define the variable that will be plotted on
the horizontal axis.}

\item{y}{The "y" parameter would define the variable that will be plotted on
the vertical axis.}

\item{xlab}{The "xlab" parameter allows you to specify the label or title for
the x-axis in your plot. The x-axis label is essential for effectively
communicating the meaning of the data represented on the horizontal axis.}

\item{ylab}{The "ylab" parameter allows you to specify the label or title for
the y-axis in your plot. The y-axis label is crucial for effectively
communicating the meaning of the data represented on the vertical axis.}

\item{shape}{The "shape" parameter is used to specify the shape of the points
in a scatter plot or other point-based plots. Points are the individual data
markers used to represent the data values in the plot.}

\item{colour}{The "colour" parameter is used to specify the colour of the
jitters in a jitter plot.}

\item{title}{The "title" parameter is used to specify the title of the plot.
The title provides a clear and concise description of the plot's purpose or
the main idea it conveys. It is typically placed above the plot to provide
context and help users understand the data being visualised.}

\item{size}{The "size" parameter is used to specify the size of the points in
a scatter plot or other point-based plots. Points are the individual data
markers used to represent the data values in the plot.}
}
\description{
Creates a jitter plot with different shapes using ggplot2.
}
\examples{
jitter_plot_shape(data = oktoberfest_daten, x = bier_preis, y = bier_konsum,
            xlab = "beer price", ylab = "beer consumption",
            shape = "circle", colour = "orange",
            title = "Relationship between beer consumption and beer price",
            subtitle = "depending on the number of visitors per day")
}
