% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisPlot.R
\name{line_chart}
\alias{line_chart}
\title{Creates a line chart using ggplot2.
Visualise trends and changes in data over continuous intervals, perfect for
time-series data or any dataset with a natural ordering.}
\usage{
line_chart(
  data,
  x,
  y,
  xlab = "x-axis",
  ylab = "y-axis",
  colour = "blue",
  title = "Line chart"
)
}
\arguments{
\item{data}{The data frame who contains the data.}

\item{x}{The variable to be shown on the x-axis.}

\item{y}{The variable to be shown on the y-axis.}

\item{xlab}{The label for the x-axis.}

\item{ylab}{The label for the y-axis.}

\item{colour}{The colour for the line.}

\item{title}{The title of the plot.}
}
\description{
Creates a line chart using ggplot2.
Visualise trends and changes in data over continuous intervals, perfect for
time-series data or any dataset with a natural ordering.
}
\examples{
line_chart(data = oktoberfest_daten, x = jahr, y = bier_konsum,
            xlab = "year", ylab = "beer consumption (in hl)",
            colour = "blue",
            title = "beer consumption during the Oktoberfest")
}
