% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisPlot.R
\name{line_chart}
\alias{line_chart}
\title{Line Chart}
\usage{
line_chart(
  data = NULL,
  x,
  y,
  xlab = "x-axis",
  ylab = "y-axis",
  colour = "blue",
  title = "Line Chart"
)
}
\arguments{
\item{data}{A data.frame or any other object provided will be used to replace
the default plot data. All objects will be transformed into a fortified
data.frame. You can refer to the fortify() function to understand which
variables will be created. Alternatively you can use a custom function that
takes the plot data as a single argument. The function must return a
data.frame, which will be utilised as the layer data.}

\item{x}{The "x" parameter defines the variable whose values will be used to
determine the position of the data points along the x-axis. The points are
then connected by lines to form the chart. The x-axis displays the values of
the x variable at different data points, allowing us to visualise how the
values of the y variable (on the y-axis) change with respect to the x
variable.}

\item{y}{The "y" parameter defines the variable whose values will be used to
determine the height or position of the points on the y-axis, which are then
connected by lines to form the chart. The y-axis displays the values of the y
variable at different data points, allowing us to visualise how the values
change over time or another continuous variable represented on the x-axis.}

\item{xlab}{The "xlab" parameter allows you to specify the label or title for
the x-axis in your plot. The x-axis label is essential for effectively
communicating the meaning of the data represented on the horizontal axis.}

\item{ylab}{The "ylab" parameter allows you to specify the label or title for
the y-axis in your plot. The y-axis label is crucial for effectively
communicating the meaning of the data represented on the vertical axis.}

\item{colour}{The "colour" parameter is used to specify the color of the
lines that connect the data points. Each line corresponds to a different
group or category in the data, and the colour parameter allows you to
customise the color of each line.}

\item{title}{The "title" parameter is used to specify the title of the plot.
The title provides a clear and concise description of the plot's purpose or
the main idea it conveys. It is typically placed above the plot to provide
context and help users understand the data being visualised.}
}
\description{
Creates a line chart using ggplot2.
Visualise trends and changes in data over continuous intervals, perfect for
time-series data or any dataset with a natural ordering.
}
\examples{
line_chart(data = oktoberfest_daten, x = jahr, y = bier_konsum,
            xlab = "year", ylab = "beer consumption (in hl)",
            colour = "blue",
            title = "beer consumption during the Oktoberfest")
}
