% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisPlot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Simplifize the visualization of plots. In the past to semester we did have to
plot many plots and we thought that it would be easier, if users would know
exactly how they have to plot something. The package visplots is there to
minimalize the work to write a good plot.}
\usage{
scatter_plot(
  data,
  x,
  y,
  xlab = "x-axis",
  ylab = "y-axis",
  colour = "purple",
  size = 5,
  title = "Scatter Plot"
)
}
\arguments{
\item{data}{The data frame who contains the data.}

\item{x}{The variable to be shown on the x-axis.}

\item{y}{The variable to be shown on the y-axis.}

\item{xlab}{The label for the x-axis.}

\item{ylab}{The label for the y-axis.}

\item{colour}{The colour for the points.}

\item{size}{The size of the points.}

\item{title}{The title of the plot.}
}
\description{
Creates a scatter plot using ggplot2.
}
\examples{
scatter_plot(data = oktoberfest_daten, x = jahr, y = besucher_gesamt,
            xlab = "Jahr", ylab = "Anzahl der Besucher (in Millionen)",
            colour = "purple", size = 3,
            title = "Anzahl der Besucher der Oktoberfeste pro Jahr")
}
