% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisPlot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Simplifize the visualization of plots. In the past two semesters we used a lot of
plots and we thought that it would be easier, if users would get information 
on how to plot the right way. The package visplots enables an reduced workload.}
\usage{
scatter_plot(
  data,
  x,
  y,
  xlab = "x-axis",
  ylab = "y-axis",
  colour = "purple",
  size = 5,
  title = "Scatter Plot"
)
}
\arguments{
\item{data}{The data frame who contains the data.}

\item{x}{The variable to be shown on the x-axis.}

\item{y}{The variable to be shown on the y-axis.}

\item{xlab}{The label for the x-axis.}

\item{ylab}{The label for the y-axis.}

\item{colour}{The colour for the points.}

\item{size}{The size of the points.}

\item{title}{The title of the plot.}
}
\description{
Creates a scatter plot using ggplot2.
}
\examples{
scatter_plot(data = oktoberfest_daten, x = jahr, y = besucher_gesamt,
            xlab = "year", ylab = "number of visitors (in millions)",
            colour = "purple", size = 3,
            title = "number of visitors of the Oktoberfest per year")
}
